+++ vck-openid-ktor/src/commonMain/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClient.kt
@@ -9,10 +9,8 @@ import at.asitplus.openid.OpenIdConstants.PARAMETER_PROMPT_LOGIN
 import at.asitplus.openid.OpenIdConstants.PATH_WELL_KNOWN_OAUTH_AUTHORIZATION_SERVER
 import at.asitplus.openid.OpenIdConstants.PATH_WELL_KNOWN_OPENID_CONFIGURATION
 import at.asitplus.openid.OpenIdConstants.TOKEN_TYPE_DPOP
-import at.asitplus.signum.indispensable.SignatureAlgorithm
 import at.asitplus.signum.indispensable.josef.JsonWebToken
 import at.asitplus.signum.indispensable.josef.JwsAlgorithm
-import at.asitplus.signum.indispensable.josef.toJwsAlgorithm
 import at.asitplus.wallet.lib.agent.DefaultCryptoService
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
 import at.asitplus.wallet.lib.agent.Holder
@@ -51,7 +49,7 @@ import kotlin.time.Duration.Companion.minutes
 /**
  * Implements the client side of
  * [OpenID for Verifiable Credential Issuance](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html)
- *  * , Draft 15, 2024-12-19.
+ *  Draft 15, 2024-12-19.
  *
  * Supported features:
  *  * Pre-authorized grants
@@ -62,7 +60,8 @@ import kotlin.time.Duration.Companion.minutes
  */
 class OpenId4VciClient(
     /** Used to continue authentication in a web browser, be sure to call back this service at [resumeWithAuthCode]. */
-    private val openUrlExternally: suspend (String) -> Unit,
+    @Deprecated("Use methods returning result instead")
+    private val openUrlExternally: (suspend (String) -> Unit)? = null,
     /** ktor engine to use to make requests to issuing service. */
     engine: HttpClientEngine,
     /**
@@ -73,9 +72,11 @@ class OpenId4VciClient(
     /** Additional configuration for building the HTTP client, e.g. callers may enable logging. */
     httpClientConfig: (HttpClientConfig<*>.() -> Unit)? = null,
     /** Store context before jumping to an external browser with [openUrlExternally]. */
-    private val storeProvisioningContext: suspend (ProvisioningContext) -> Unit,
+    @Deprecated("Use methods returning result instead")
+    private val storeProvisioningContext: (suspend (ProvisioningContext) -> Unit)? = null,
     /** Load context after resuming with auth code in [resumeWithAuthCode]. */
-    private val loadProvisioningContext: suspend () -> ProvisioningContext?,
+    @Deprecated("Use methods returning result instead")
+    private val loadProvisioningContext: (suspend () -> ProvisioningContext?)? = null,
     /**
      * Callback to load the client attestation JWT, which may be needed as authentication at the AS, where the
      * `clientId` must match [WalletService.clientId] in [oid4vciService] and the key attested in `cnf` must match
@@ -86,7 +87,10 @@ class OpenId4VciClient(
     @Deprecated("Use signClientAttestationPop instead")
     private val clientAttestationJwsService: JwsService = DefaultJwsService(DefaultCryptoService(EphemeralKeyWithoutCert())),
     /** Used for authenticating the client at the authorization server with client attestation. */
-    private val signClientAttestationPop: SignJwtFun<JsonWebToken> = SignJwt(EphemeralKeyWithoutCert(), JwsHeaderNone()),
+    private val signClientAttestationPop: SignJwtFun<JsonWebToken> = SignJwt(
+        EphemeralKeyWithoutCert(),
+        JwsHeaderNone()
+    ),
     @Deprecated("Use signDpop instead")
     private val dpopJwsService: JwsService = DefaultJwsService(DefaultCryptoService(EphemeralKeyWithoutCert())),
     /** Used to calculate DPoP, i.e. the key the access token and refresh token gets bound to. */
@@ -98,9 +102,11 @@ class OpenId4VciClient(
      */
     val oid4vciService: WalletService = WalletService(),
     /** Final callback upon receiving credentials from the issuing service. */
-    private val storeCredential: (suspend (Holder.StoreCredentialInput) -> Unit),
-    /** Callback to store refresh tokens received from the AS, to refresh credentials sometime late. */
-    private val storeRefreshToken: suspend (RefreshTokenInfo) -> Unit,
+    @Deprecated("Use methods returning result instead")
+    private val storeCredential: (suspend (Holder.StoreCredentialInput) -> Unit)? = null,
+    /** Callback to store refresh tokens received from the AS, to refresh credentials sometime later. */
+    @Deprecated("Use methods returning result instead")
+    private val storeRefreshToken: (suspend (RefreshTokenInfo) -> Unit)? = null,
 ) {
     private val client: HttpClient = HttpClient(engine) {
         followRedirects = false
@@ -157,10 +163,34 @@ class OpenId4VciClient(
      * @param credentialIssuerUrl URL of the credential issuer service
      * @param credentialIdentifierInfo credential to request, i.e. picked by user selection
      */
+    @Deprecated(
+        "Use startProvisioningWithAuthRequestReturningResult instead",
+        ReplaceWith("startProvisioningWithAuthRequestReturningResult(credentialIssuerUrl, credentialIdentifierInfo)")
+    )
     suspend fun startProvisioningWithAuthRequest(
         credentialIssuerUrl: String,
         credentialIdentifierInfo: CredentialIdentifierInfo,
     ): KmmResult<Unit> = catching {
+        startProvisioningWithAuthRequestReturningResult(credentialIssuerUrl, credentialIdentifierInfo).onSuccess {
+            storeProvisioningContext?.invoke(it.context)
+            openUrlExternally?.invoke(it.url)
+        }
+    }
+
+    /**
+     * Starts the issuing process at [credentialIssuerUrl].
+     * Clients need to handle the result, i.e. open the URL for user authentication or store the credentials.
+     * Clients need to call [resumeWithAuthCode] after getting the authorization code back from the authorization
+     * server, e.g. by the Wallet app getting opened (see `redirectUrl` at [oid4vciService]) after the browser being
+     * redirecting back from the authorization server.
+     *
+     * @param credentialIssuerUrl URL of the credential issuer service
+     * @param credentialIdentifierInfo credential to request, i.e. picked by user selection
+     */
+    suspend fun startProvisioningWithAuthRequestReturningResult(
+        credentialIssuerUrl: String,
+        credentialIdentifierInfo: CredentialIdentifierInfo,
+    ): KmmResult<CredentialIssuanceResult.OpenUrlForAuthnRequest> = catching {
         Napier.i("startProvisioningWithAuthRequest: $credentialIssuerUrl with $credentialIdentifierInfo")
 
         val issuerMetadata = credentialIdentifierInfo.issuerMetadata
@@ -175,26 +205,12 @@ class OpenId4VciClient(
         }
 
         val state = uuid4().toString()
-        ProvisioningContext(
+        startAuthorization(
             state = state,
-            credential = credentialIdentifierInfo,
-            oauthMetadata = oauthMetadata,
-            issuerMetadata = issuerMetadata
-        ).let {
-            storeProvisioningContext.invoke(it)
-            Napier.i("Store context: $it")
-        }
-
-        openAuthRequestInBrowser(
-            state = state,
-            scope = credentialIdentifierInfo.supportedCredentialFormat.scope,
-            authorizationDetails = oid4vciService.buildAuthorizationDetails(
-                credentialIdentifierInfo.credentialIdentifier,
-                issuerMetadata.authorizationServers
-            ),
+            credentialIdentifierInfo = credentialIdentifierInfo,
+            issuerMetadata = issuerMetadata,
             credentialIssuer = credentialIssuerUrl,
-            issuerState = null,
-            oauthMetadata = oauthMetadata
+            oauthMetadata = oauthMetadata,
         )
     }
 
@@ -209,14 +225,37 @@ class OpenId4VciClient(
      *
      * @param url the URL as it has been redirected back from the authorization server, i.e. containing param `code`
      */
-    @Throws(Exception::class)
+    @Suppress("DEPRECATION")
+    @Deprecated("Use resumeWithAuthCode with context instead", ReplaceWith("resumeWithAuthCode(url, context)"))
     suspend fun resumeWithAuthCode(
         url: String,
     ): KmmResult<Unit> = catching {
         Napier.i("resumeWithAuthCode")
-        Napier.d("resumeWithAuthCode: $url")
-        val context = loadProvisioningContext()
+        val context = loadProvisioningContext?.invoke()
             ?: throw Exception("No provisioning context")
+        resumeWithAuthCode(url, context).onSuccess {
+            it.credentials.forEach { storeCredential?.invoke(it) }
+            it.refreshToken?.let { storeRefreshToken?.invoke(it) }
+        }.getOrThrow()
+    }
+
+    /**
+     * Called after getting the redirect back from the authorization server to the credential issuer.
+     *
+     * Will request a token, and use that token to request a credential and store it.
+     *
+     * Prefers building the token request by using `scope` (from [SupportedCredentialFormat]), as advised in
+     * [OpenID4VC HAIP](https://openid.net/specs/openid4vc-high-assurance-interoperability-profile-1_0.html),
+     * but falls back to authorization details if needed.
+     *
+     * @param url the URL as it has been redirected back from the authorization server, i.e. containing param `code`
+     */
+    suspend fun resumeWithAuthCode(
+        url: String,
+        context: ProvisioningContext,
+    ): KmmResult<CredentialIssuanceResult.Success> = catching {
+        Napier.i("resumeWithAuthCode")
+        Napier.d("resumeWithAuthCode: $url, $context")
 
         val authnResponse = Url(url).parameters.flattenEntries().toMap()
             .decodeFromUrlQuery<AuthenticationResponseParameters>()
@@ -264,10 +303,33 @@ class OpenId4VciClient(
      * [OpenID4VC HAIP](https://openid.net/specs/openid4vc-high-assurance-interoperability-profile-1_0.html),
      * but falls back to authorization details if needed.
      */
-    @Throws(Exception::class)
+    @Suppress("DEPRECATION")
+    @Deprecated(
+        "Use refreshCredentialReturningResult instead",
+        ReplaceWith("refreshCredentialReturningResult(refreshTokenInfo)")
+    )
     suspend fun refreshCredential(
         refreshTokenInfo: RefreshTokenInfo,
     ): KmmResult<Unit> = catching {
+        refreshCredentialReturningResult(refreshTokenInfo).onSuccess {
+            it.credentials.forEach { storeCredential?.invoke(it) }
+            it.refreshToken?.let { storeRefreshToken?.invoke(it) }
+        }.getOrThrow()
+    }
+
+    /**
+     * Call to refresh a credential with a stored refresh token (that was received when issuing the credential
+     * for the first time, stored with [storeRefreshToken]).
+     *
+     * Will request a new access token, and use that token to request the same credential again and store it.
+     *
+     * Prefers building the token request by using `scope` (from [SupportedCredentialFormat]), as advised in
+     * [OpenID4VC HAIP](https://openid.net/specs/openid4vc-high-assurance-interoperability-profile-1_0.html),
+     * but falls back to authorization details if needed.
+     */
+    suspend fun refreshCredentialReturningResult(
+        refreshTokenInfo: RefreshTokenInfo,
+    ): KmmResult<CredentialIssuanceResult.Success> = catching {
         with(refreshTokenInfo) {
             Napier.i("refreshCredential")
             Napier.d("refreshCredential: $refreshToken, $credentialFormat, $credentialIdentifier")
@@ -359,8 +421,8 @@ class OpenId4VciClient(
         credentialScheme: ConstantIndex.CredentialScheme,
         oauthMetadata: OAuth2AuthorizationServerMetadata,
         credentialIdentifier: String,
-        previouslyRequestedScope: String?
-    ) {
+        previouslyRequestedScope: String?,
+    ): CredentialIssuanceResult.Success {
         val credentialEndpointUrl = issuerMetadata.credentialEndpointUrl
         Napier.i("postCredentialRequestAndStore: $credentialEndpointUrl")
         Napier.d("postCredentialRequestAndStore: $tokenResponse")
@@ -383,19 +445,7 @@ class OpenId4VciClient(
             BuildDPoPHeader(signDpop, url = credentialEndpointUrl, accessToken = tokenResponse.accessToken)
         else null
 
-        if (tokenResponse.refreshToken != null) {
-            storeRefreshToken.invoke(
-                RefreshTokenInfo(
-                    refreshToken = tokenResponse.refreshToken!!,
-                    issuerMetadata = issuerMetadata,
-                    oauthMetadata = oauthMetadata,
-                    credentialFormat = credentialFormat,
-                    credentialIdentifier = credentialIdentifier,
-                )
-            )
-        }
-
-        credentialRequests.forEach { credentialRequest ->
+        val storeCredentialInputs = credentialRequests.flatMap { credentialRequest ->
             val credentialResponse: CredentialResponseParameters = client.post(credentialEndpointUrl) {
                 contentType(ContentType.Application.Json)
                 setBody(credentialRequest)
@@ -407,12 +457,19 @@ class OpenId4VciClient(
 
             credentialResponse.extractCredentials()
                 .ifEmpty { throw Exception("No credential was received") }
-                .forEach {
-                    storeCredential.invoke(
-                        it.toStoreCredentialInput(credentialFormat.format.toRepresentation(), credentialScheme)
-                    )
-                }
+                .map { it.toStoreCredentialInput(credentialFormat.format.toRepresentation(), credentialScheme) }
         }
+        return CredentialIssuanceResult.Success(
+            storeCredentialInputs,
+            tokenResponse.refreshToken?.let {
+                RefreshTokenInfo(
+                    refreshToken = tokenResponse.refreshToken!!,
+                    issuerMetadata = issuerMetadata,
+                    oauthMetadata = oauthMetadata,
+                    credentialFormat = credentialFormat,
+                    credentialIdentifier = credentialIdentifier,
+                )
+            })
     }
 
     /**
@@ -422,12 +479,42 @@ class OpenId4VciClient(
      * @param credentialIdentifierInfo as selected by the user from the issuer's metadata
      * @param transactionCode if required from Issuing service, i.e. transmitted out-of-band to the user
      */
-    @Throws(Throwable::class)
+    @Deprecated(
+        "Use startProvisioningWithAuthRequestReturningResult instead",
+        ReplaceWith("startProvisioningWithAuthRequestReturningResult(credentialIssuerUrl, credentialIdentifierInfo, transactionCode)")
+    )
     suspend fun loadCredentialWithOffer(
         credentialOffer: CredentialOffer,
         credentialIdentifierInfo: CredentialIdentifierInfo,
         transactionCode: String? = null,
     ): KmmResult<Unit> = catching {
+        loadCredentialWithOfferReturningResult(credentialOffer, credentialIdentifierInfo, transactionCode).onSuccess {
+            when (it) {
+                is CredentialIssuanceResult.OpenUrlForAuthnRequest -> {
+                    storeProvisioningContext?.invoke(it.context)
+                    openUrlExternally?.invoke(it.url)
+                }
+
+                is CredentialIssuanceResult.Success -> {
+                    it.credentials.forEach { storeCredential?.invoke(it) }
+                    it.refreshToken?.let { storeRefreshToken?.invoke(it) }
+                }
+            }
+        }.getOrThrow()
+    }
+
+    /**
+     * Loads a user-selected credential with pre-authorized code from the OID4VCI credential issuer
+     *
+     * @param credentialOffer as loaded and decoded from the QR Code
+     * @param credentialIdentifierInfo as selected by the user from the issuer's metadata
+     * @param transactionCode if required from Issuing service, i.e. transmitted out-of-band to the user
+     */
+    suspend fun loadCredentialWithOfferReturningResult(
+        credentialOffer: CredentialOffer,
+        credentialIdentifierInfo: CredentialIdentifierInfo,
+        transactionCode: String? = null,
+    ): KmmResult<CredentialIssuanceResult> = catching {
         Napier.i("loadCredentialWithOffer: $credentialOffer")
         val issuerMetadata = credentialIdentifierInfo.issuerMetadata
         val authorizationServer = issuerMetadata.authorizationServers?.firstOrNull()
@@ -472,23 +559,10 @@ class OpenId4VciClient(
                 previouslyRequestedScope = credentialIdentifierInfo.supportedCredentialFormat.scope,
             )
         } ?: credentialOffer.grants?.authorizationCode?.let {
-            ProvisioningContext(
+            startAuthorization(
                 state = state,
-                credential = credentialIdentifierInfo,
-                oauthMetadata = oauthMetadata,
-                issuerMetadata = issuerMetadata
-            ).let {
-                storeProvisioningContext.invoke(it)
-                Napier.d("Store context: $it")
-            }
-
-            openAuthRequestInBrowser(
-                state = state,
-                scope = credentialIdentifierInfo.supportedCredentialFormat.scope,
-                authorizationDetails = oid4vciService.buildAuthorizationDetails(
-                    credentialIdentifierInfo.credentialIdentifier,
-                    issuerMetadata.authorizationServers
-                ),
+                credentialIdentifierInfo = credentialIdentifierInfo,
+                issuerMetadata = issuerMetadata,
                 credentialIssuer = credentialOffer.credentialIssuer,
                 issuerState = it.issuerState,
                 oauthMetadata = oauthMetadata,
@@ -519,19 +593,22 @@ class OpenId4VciClient(
      * Uses Pushed Authorization Requests [RFC 9126](https://datatracker.ietf.org/doc/html/rfc9126) if advised
      * by the authorization server.
      *
-     * Will call [openUrlExternally] in the end to perform user authentication in a browser.
-     *
-     * Process continues at [resumeWithAuthCode].
+     * Clients need to contiune the process (after getting back from the browser) with [resumeWithAuthCode].
      */
     @Throws(Exception::class)
-    private suspend fun openAuthRequestInBrowser(
+    private suspend fun startAuthorization(
         state: String,
-        scope: String?,
-        authorizationDetails: Set<OpenIdAuthorizationDetails>,
+        credentialIdentifierInfo: CredentialIdentifierInfo,
+        issuerMetadata: IssuerMetadata,
         credentialIssuer: String,
         issuerState: String? = null,
         oauthMetadata: OAuth2AuthorizationServerMetadata,
-    ) {
+    ): CredentialIssuanceResult.OpenUrlForAuthnRequest {
+        val scope = credentialIdentifierInfo.supportedCredentialFormat.scope
+        val authorizationDetails = oid4vciService.buildAuthorizationDetails(
+            credentialIdentifierInfo.credentialIdentifier,
+            issuerMetadata.authorizationServers
+        )
         val authorizationEndpointUrl = oauthMetadata.authorizationEndpoint
             ?: throw Exception("no authorizationEndpoint in $oauthMetadata")
         val authRequest = oid4vciService.oauth2Client.createAuthRequest(
@@ -563,8 +640,14 @@ class OpenId4VciClient(
                 builder.parameters.append(PARAMETER_PROMPT, PARAMETER_PROMPT_LOGIN)
             }.build().toString()
         }
-        Napier.i("Provisioning starts by opening URL $authorizationUrl")
-        openUrlExternally.invoke(authorizationUrl)
+        val context = ProvisioningContext(
+            state = state,
+            credential = credentialIdentifierInfo,
+            oauthMetadata = oauthMetadata,
+            issuerMetadata = issuerMetadata
+        )
+        Napier.i("Provisioning starts by returning URL to open: $authorizationUrl")
+        return CredentialIssuanceResult.OpenUrlForAuthnRequest(authorizationUrl, context)
     }
 
     @Throws(Exception::class)
@@ -629,6 +712,29 @@ data class ProvisioningContext(
     val oauthMetadata: OAuth2AuthorizationServerMetadata,
     val issuerMetadata: IssuerMetadata,
 )
+
+/**
+ * Result of the credential issuance process: Either open an authentication request URL externally (i.e. the browser),
+ * or store the received credentials.
+ */
+sealed interface CredentialIssuanceResult {
+    /**
+     * Store credentials in [credentials], and optionally the [refreshToken] for a later renewal of those credentials.
+     */
+    data class Success(
+        val credentials: Collection<Holder.StoreCredentialInput>,
+        val refreshToken: RefreshTokenInfo? = null,
+    ) : CredentialIssuanceResult
+
+    /**
+     * Open the [url] in a browser (so the user can authenticate at the AS), and store [context] to use in next call
+     * to [at.asitplus.wallet.lib.ktor.openid.OpenId4VciClient.resumeWithAuthCode].
+     */
+    data class OpenUrlForAuthnRequest(
+        val url: String,
+        val context: ProvisioningContext,
+    ) : CredentialIssuanceResult
+}
 
 /**
  * Gets parsed from the credential issuer's metadata, essentially an entry from